+++ CHANGES.rst
@@ -18,24 +18,45 @@ Release History
    - Removed
    - Fixed
 
+<<<<<<< HEAD
 0.5.0 (unreleased)
+=======
+0.5.1 (unreleased)
+>>>>>>> master
 ------------------
+
+**Changed**
+
+- ``sim.data[obj]`` will now return live parameter values from the simulation,
+  rather than initial values from the build process.  That means that it can
+  be used to get the values of object parameters after training, e.g.
+  ``sim.data[my_conn].weights``.
+
+
+0.5.0 (July 11, 2017)
+---------------------
 
 **Added**
 
 - Added ``nengo_dl.tensor_layer`` to help with the construction of
   layer-style TensorNodes (see the `TensorNode documentation
-  <https://nengo.github.io/nengo_dl/tensor_node.html>`_)
+  <http://www.nengo.ai/nengo_dl/tensor_node.html>`_)
 - Added an example demonstrating `how to train a neural network
   that can run in spiking neurons
-  <https://nengo.github.io/nengo_dl/examples/spiking_mnist.html>`_
+  <http://www.nengo.ai/nengo_dl/examples/spiking_mnist.html>`_
 - Added some distributions for weight initialization to ``nengo_dl.dists``
-- Added ``sim.train(..., profile=True)` option to collect profiling information
+- Added ``sim.train(..., profile=True)`` option to collect profiling information
   during training
 - Added new methods to simplify the Nengo operation graph, resulting in faster
   simulation/training speed
 - The default graph planner can now be modified by setting the ``planner``
   attribute on the top-level Network config
+<<<<<<< HEAD
+=======
+- Added TensorFlow implementation for general linear synapses
+- Added ``backports.tempfile`` and ``backports.print_function`` requirement for
+  Python 2.7 systems
+>>>>>>> master
 
 **Changed**
 
@@ -47,8 +68,16 @@ Release History
   the more general ``nengo_dl.configure_settings(trainable=x)``.  This has
   resulted in some small changes to how trainability is controlled within
   subnetworks; see the `updated documentation
+<<<<<<< HEAD
   <https://nengo.github.io/nengo_dl/training.html#choosing-which-elements-to-optimize>`_
   for details.
+=======
+  <http://www.nengo.ai/nengo_dl/training.html#choosing-which-elements-to-optimize>`_
+  for details.
+- Calling ``Simulator.train``/``Simulator.loss`` no longer resets the internal
+  state of the simulation (so they can be safely intermixed with calls to
+  ``Simulator.run``)
+>>>>>>> master
 
 **Deprecated**
 
@@ -58,7 +87,7 @@ Release History
 **Fixed**
 
 - Fixed bug related to changing the output of a Node after the model is
-  constructed (`#4 <https://github.com/nengo/nengo_dl/issues/4)>`_)
+  constructed (`#4 <https://github.com/nengo/nengo_dl/issues/4>`_)
 - Order of variable creation is now deterministic (helps make saving/loading
   parameters more reliable)
 - Configuring whether or not a model element is trainable does not affect
@@ -68,6 +97,14 @@ Release History
 - Correctly handle probes that aren't connected to any ops
 - Swapped ``fan_in``/``fan_out`` in ``dists.VarianceScaling`` to align with
   the standard definitions
+<<<<<<< HEAD
+=======
+- Temporary patch to fix memory leak in TensorFlow (see 
+  `#11273 <https://github.com/tensorflow/tensorflow/issues/11273>`_)
+- Fixed bug related to nodes that had matching output functions but different
+  size_out
+- Fixed bug related to probes that do not contain any data yet
+>>>>>>> master
 
 0.4.0 (June 8, 2017)
 --------------------
@@ -76,10 +113,10 @@ Release History
 
 - Added ability to manually specify which parts of a model are trainable
   (see the `sim.train documentation
-  <https://nengo.github.io/nengo_dl/training.html>`_)
+  <http://www.nengo.ai/nengo_dl/training.html>`_)
 - Added some code examples (see the ``docs/examples`` directory, or the
   `pre-built examples in the documentation
-  <https://nengo.github.io/nengo_dl/examples.html>`_)
+  <http://www.nengo.ai/nengo_dl/examples.html>`_)
 - Added the SoftLIFRate neuron type for training LIF networks (based on
   `this paper <https://arxiv.org/abs/1510.08829>`_)